requirements.txt
discord.py >= 2.5.2
yt-dlp >= 2025.3.27
PyNaCl >= 1.5.0
ffmpeg-python >= 0.2.0